+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Luminati Proxy manager - Change Log
 
+## 1.142.102 Stable
+- :star: Improved performance when using a big number of proxy ports
+- :bug: Fixed zones/plans details
+
 ## 1.141.733 Stable
 - :star: Added api/async_req endpoint